# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Deployer API",
    "description": "REST API for deploying containers from templates. Async: POST deploy/lifecycle → 202 + operation, poll GET /api/operations/:id. Deploy: containerName + params.",
    "version": "2.0.0"
  },
  "servers": [{ "url": "/", "description": "Current host" }],
  "components": {
    "securitySchemes": {
      "session": { "type": "apiKey", "in": "cookie", "name": "s" },
      "apiKey": { "type": "apiKey", "in": "header", "name": "X-API-Key" },
      "mcpBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "MCP key dep_mcp_live_… (full Deployer API via /mcp tools)"
      }
    },
    "schemas": {
      "ContainerRef": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "state": { "type": "string" }
        }
      },
...
      "DeployRequest": {
        "type": "object",
        "required": ["templateId", "containerName", "params"],
        "properties": {
          "templateId": { "type": "string", "example": "docker-demo-free" },
          "containerName": {
            "type": "string",
            "description": "Docker container name. Set at deploy time, not in template JSON."
          },
          "params": {
            "type": "object",
            "additionalProperties": true,
            "description": "Template form fields (HOST_PORT, DOMAIN, …). Excludes container name."
          }
        }
      },
...
      "AcceptedOperation": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "operation": { "$ref": "#/components/schemas/Operation" }
        }
      }
    }
  },
  "security": [{ "session": [] }, { "apiKey": [] }],
  "paths": {
    "/api/login": {
      "post": {
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["username", "password"], "properties": { "username": { "type": "string" }, "password": { "type": "string" } } } } }
        },
        "responses": { "200": { "description": "OK, session established" }, "401": { "description": "Invalid credentials" } },
        "security": []
      }
    },
...
    "/api/me": {
      "get": {
        "summary": "Current user",
        "responses": { "200": { "description": "user" }, "401": { "description": "Not authenticated" } }
      }
    },
...
    "/mcp": {
      "post": {
        "summary": "MCP JSON-RPC (Bearer dep_mcp_live_…)",
        "description": "Streamable HTTP MCP. See docs/DEPLOYER-MCP-v1-RU.md",
        "security": [{ "mcpBearer": [] }],
        "responses": { "200": { "description": "JSON-RPC result" }, "401": { "description": "Invalid key" } }
      }
    },
...
    "/api/templates": {
      "get": {
        "summary": "List templates",
        "responses": { "200": { "description": "Array of templates" } }
      },
      "post": {
        "summary": "Create or update template",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Template" } } } },
        "responses": { "200": { "description": "Saved template" }, "400": { "description": "Validation error" } }
      }
    },
    "/api/templates/{id}": {
      "get": {
        "summary": "Get template by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Template" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "summary": "Delete template",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" }, "400": { "description": "Error" } }
      }
    },
    "/api/deploy": {
      "post": {
        "summary": "Deploy container from template (async)",
        "description": "202 + operation. containerName = Docker name; params = template fields.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeployRequest" }
            }
          }
        },
        "responses": {
          "202": { "description": "Operation accepted", "headers": { "Location": { "schema": { "type": "string" } }, "Retry-After": { "schema": { "type": "integer" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AcceptedOperation" } } } },
          "200": { "description": "Sync mode (DEPLOYER_SYNC_LEGACY=1): container in body", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" }, "container": { "$ref": "#/components/schemas/ContainerRef" }, "params": { "type": "object" } } } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Not authenticated" },
          "404": { "description": "Template not found" },
          "409": { "description": "operation_in_progress", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OperationConflict" } } } },
          "429": { "description": "Rate limit" },
          "503": { "description": "deployer_busy", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeployerBusy" } } } },
          "500": { "description": "Error Docker" }
        }
      }
    },
    "/api/operations/{id}": {
      "get": {
        "summary": "Async operation status",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "operation + Retry-After: 2", "headers": { "Retry-After": { "schema": { "type": "integer" } } }, "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" }, "operation": { "$ref": "#/components/schemas/Operation" } } } } } },
          "404": { "description": "operation_not_found (e.g. after Deployer restart)" },
          "429": { "description": "Rate limit poll" }
        }
      }
    },
    "/api/containers": {
      "get": {
        "summary": "List containers (pagination + deploy limit)",
        "parameters": [
          { "name": "all", "in": "query", "schema": { "type": "string" }, "description": "true — include stopped" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Page size (default 20, max 100)" },
          { "name": "offset", "in": "query", "schema": { "type": "integer" }, "description": "Page offset" },
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "Search by name/image/deployName" }
        ],
        "responses": {
          "200": {
            "description": "containers[] page; total, offset, limit, page, total_pages, has_more; container_limit (0 = unlimited)"
          }
        }
      }
    },
    "/api/containers/{id}": {
      "get": {
        "summary": "Container details",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Container ID or name" },
          { "name": "stats", "in": "query", "schema": { "type": "string" }, "description": "If 1 — include stats (memory, CPU)" }
        ],
        "responses": { "200": { "description": "Container data (optional stats)" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "summary": "Stop and delete container (async); purge deploy data dir when removeData=true",
        "description": "Path :id — same as other lifecycle routes: deploy containerName or Docker container id while it exists. removeData=true deletes DEPLOY_BASE_PATH/<deploy-name>/ even when the container is already gone (202, not 404). templateId query selects deprovision script; overrides deployer.templateId label. deprovision runs only when removeData=true.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Deploy containerName or Docker container id (same as GET/POST lifecycle :id)" },
          { "name": "removeData", "in": "query", "schema": { "type": "boolean", "default": false }, "description": "Also delete host data under DEPLOY_BASE_PATH/<containerName>/ and run template deprovision when template resolves" },
          { "name": "templateId", "in": "query", "schema": { "type": "string" }, "description": "Template for deprovision; takes priority over deployer.templateId label on container" }
        ],
        "responses": {
          "202": { "description": "Operation accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AcceptedOperation" } } } },
          "200": { "description": "Sync mode (DEPLOYER_SYNC_LEGACY=1)" },
          "404": { "description": "Template not found when deprovision is required (before data removal)" },
          "409": { "description": "operation_in_progress" },
          "503": { "description": "queue_full" },
          "429": { "description": "Rate limit" }
        }
      }
    },
    "/api/containers/{id}/stats": {
      "get": {
        "summary": "Container stats (state, memory, CPU)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "state, memory_usage_mb, memory_limit_mb, cpu_percent (when running)" }, "404": { "description": "Not found" } }
      }
    },
    "/api/containers/{id}/disk": {
      "get": {
        "summary": "Container disk usage",
        "description": "Mounted dir sizes (du in container) + writable layer. Limit: 60 req/min per IP.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "state, data_bytes, writable_bytes, total_bytes, total_mb" },
          "404": { "description": "Container not found" },
          "429": { "description": "Too many requests" }
        }
      }
    },
    "/api/containers/{id}/logs": {
      "get": {
        "summary": "Container logs (docker logs)",
        "description": "Last stdout/stderr lines. Limit 60/min. tail 1–5000, default 500.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "tail", "in": "query", "schema": { "type": "integer", "default": 500, "minimum": 1, "maximum": 5000 } },
          { "name": "timestamps", "in": "query", "schema": { "type": "boolean", "default": false }, "description": "Docker timestamp prefix (--timestamps)" }
        ],
        "responses": {
          "200": { "description": "id, name, state, tail, logs" },
          "404": { "description": "Container not found" },
          "429": { "description": "Too many requests" }
        }
      }
    },
    "/api/containers/{id}/restart": {
      "post": {
        "summary": "Restart container (async)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "202": { "description": "Operation accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AcceptedOperation" } } } },
          "200": { "description": "Sync mode (DEPLOYER_SYNC_LEGACY=1)" },
          "404": { "description": "Not found" },
          "409": { "description": "operation_in_progress" },
          "503": { "description": "queue_full" },
          "429": { "description": "Rate limit" }
        }
      }
    },
    "/api/containers/{id}/stop": {
      "post": {
        "summary": "Stop container (async)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "202": { "description": "Operation accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AcceptedOperation" } } } },
          "200": { "description": "Sync mode (DEPLOYER_SYNC_LEGACY=1)" },
          "404": { "description": "Not found" },
          "409": { "description": "operation_in_progress" },
          "503": { "description": "queue_full" },
          "429": { "description": "Rate limit" }
        }
      }
    },
...
  }
}